posted the current test program
This one was originally named "fuzzytest.py", used to test the implementation of the fuzzywuzzy library for my wiki.py test file. Last time I checked, which was in English class, there were still a few tracebacks as a result of not fully configuring the logic behind some of the conditionals.