Created terms2.txt
Added the generic file listing possible terms that can be defined, as well as possible obstacles such as utter gibberish which would normally result in tracebacks. Eventually, this should be renamed both here and in the code to be something more professional like "input.txt", but hopefully by then I'll have a GUI up and running.